Refactor duplicate code from X11 and DRM backends
