+++ src/server_state.rs
@@ -4,24 +4,21 @@ use std::env::{remove_var, set_var};
 use std::rc::Rc;
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
-use std::time::Duration;
 
 use smithay::{delegate_compositor, delegate_data_device, delegate_dmabuf, delegate_output, delegate_seat, delegate_shm, delegate_xdg_shell};
 use smithay::backend::allocator::dmabuf::Dmabuf;
-use smithay::backend::input::ButtonState;
 use smithay::backend::renderer::{ImportAll, Texture};
 use smithay::backend::renderer::gles::ffi::Gles2;
 use smithay::backend::renderer::gles::GlesRenderer;
 use smithay::input::{Seat, SeatHandler, SeatState};
-use smithay::input::pointer::{ButtonEvent, CursorImageStatus, MotionEvent, PointerHandle};
-use smithay::reexports::calloop::{channel, Interest, LoopHandle, Mode, PostAction};
-use smithay::reexports::calloop::channel::Event::Msg;
+use smithay::input::pointer::{CursorImageStatus, PointerHandle};
+use smithay::reexports::calloop::{Interest, LoopHandle, Mode, PostAction};
 use smithay::reexports::calloop::generic::Generic;
 use smithay::reexports::wayland_protocols::xdg::shell::server::xdg_toplevel;
 use smithay::reexports::wayland_server::{Client, Display, DisplayHandle, Resource};
 use smithay::reexports::wayland_server::protocol::{wl_buffer, wl_seat};
 use smithay::reexports::wayland_server::protocol::wl_surface::WlSurface;
-use smithay::utils::{Buffer as BufferCoords, Clock, Monotonic, Rectangle, Serial, SERIAL_COUNTER, Size};
+use smithay::utils::{Buffer as BufferCoords, Clock, Monotonic, Rectangle, Serial, Size};
 use smithay::wayland::buffer::BufferHandler;
 use smithay::wayland::compositor;
 use smithay::wayland::compositor::{BufferAssignment, CompositorClientState, CompositorHandler, CompositorState, SubsurfaceCachedState, SurfaceAttributes, TraversalAction, with_states, with_surface_tree_upward};
@@ -38,12 +35,9 @@ use tracing::{info, warn};
 use crate::{Backend, CalloopData, ClientState};
 use crate::flutter_engine::FlutterEngine;
 use crate::flutter_engine::platform_channels::encodable_value::EncodableValue;
-use crate::flutter_engine::platform_channels::method_call::MethodCall;
 use crate::flutter_engine::platform_channels::method_channel::MethodChannel;
-use crate::flutter_engine::platform_channels::method_result::MethodResult;
 use crate::flutter_engine::platform_channels::standard_method_codec::StandardMethodCodec;
 use crate::flutter_engine::wayland_messages::{SubsurfaceCommitMessage, SurfaceCommitMessage, XdgPopupCommitMessage, XdgSurfaceCommitMessage};
-use crate::mouse_button_tracker::FLUTTER_TO_LINUX_MOUSE_BUTTONS;
 use crate::texture_swap_chain::TextureSwapChain;
 
 pub struct ServerState<BackendData: Backend + 'static> {
@@ -78,8 +72,6 @@ pub struct ServerState<BackendData: Backend + 'static> {
     pub texture_ids_per_view_id: HashMap<u64, Vec<i64>>,
     pub view_id_per_texture_id: HashMap<i64, u64>,
     pub texture_swapchains: HashMap<i64, TextureSwapChain>,
-
-    pub tx_platform_message: Option<channel::Sender<(MethodCall, Box<dyn MethodResult>)>>,
 }
 
 impl<BackendData: Backend + 'static> ServerState<BackendData> {
@@ -172,142 +164,6 @@ impl<BackendData: Backend + 'static> ServerState<BackendData> {
             )
             .expect("Failed to init wayland server source");
 
-        let (tx_platform_message, rx_platform_message) = channel::channel::<(MethodCall, Box<dyn MethodResult>)>();
-
-        macro_rules! extract {
-            ($e:expr, $p:path) => {
-                match $e {
-                    $p(value) => value,
-                    _ => panic!("Failed to extract value"),
-                }
-            };
-        }
-
-        fn get_value<'a>(map: &'a EncodableValue, key: &str) -> &'a EncodableValue {
-            let map = extract!(map, EncodableValue::Map);
-            for (k, v) in map {
-                if let EncodableValue::String(k) = k {
-                    if k == key {
-                        return v;
-                    }
-                }
-            }
-            panic!("Key {} not found in map", key);
-        }
-
-        loop_handle
-            .insert_source(
-                rx_platform_message,
-                |event, (), data| {
-                    if let Msg((method_call, mut result)) = event {
-                        let pointer = data.state.pointer.clone();
-                        let now = Duration::from(data.state.clock.now()).as_millis() as u32;
-
-                        match method_call.method() {
-                            "pointer_hover" => {
-                                let args = method_call.arguments().unwrap();
-                                let view_id = get_value(args, "view_id").long_value().unwrap();
-                                let x = *extract!(get_value(args, "x"), EncodableValue::Double);
-                                let y = *extract!(get_value(args, "y"), EncodableValue::Double);
-
-                                data.state.view_id_under_cursor = Some(view_id as u64);
-
-                                if let Some(surface) = data.state.surfaces.get(&(view_id as u64)).cloned() {
-                                    pointer.motion(
-                                        &mut data.state,
-                                        Some((surface.clone(), (0, 0).into())),
-                                        &MotionEvent {
-                                            location: (x, y).into(),
-                                            serial: SERIAL_COUNTER.next_serial(),
-                                            time: now,
-                                        },
-                                    );
-                                    pointer.frame(&mut data.state);
-                                    result.success(None);
-                                } else {
-                                    result.error(
-                                        "surface_doesnt_exist".to_string(),
-                                        format!("Surface {view_id} doesn't exist"),
-                                        None,
-                                    );
-                                }
-                            }
-                            "pointer_exit" => {
-                                data.state.view_id_under_cursor = None;
-
-                                pointer.motion(
-                                    &mut data.state,
-                                    None,
-                                    &MotionEvent {
-                                        location: (0.0, 0.0).into(),
-                                        serial: SERIAL_COUNTER.next_serial(),
-                                        time: now,
-                                    },
-                                );
-                                result.success(None);
-                            }
-                            "mouse_button_event" => {
-                                let args = method_call.arguments().unwrap();
-                                let button = get_value(args, "button").long_value().unwrap();
-                                let is_pressed = *extract!(get_value(args, "is_pressed"), EncodableValue::Bool);
-
-                                pointer.button(
-                                    &mut data.state,
-                                    &ButtonEvent {
-                                        serial: SERIAL_COUNTER.next_serial(),
-                                        time: now,
-                                        button: *FLUTTER_TO_LINUX_MOUSE_BUTTONS.get(&(button as u32)).unwrap() as u32,
-                                        state: if is_pressed { ButtonState::Pressed } else { ButtonState::Released },
-                                    },
-                                );
-                                pointer.frame(&mut data.state);
-                                result.success(None);
-                            }
-                            "activate_window" => {
-                                let args = method_call.arguments().unwrap();
-                                let args = extract!(args, EncodableValue::List);
-                                let view_id = args[0].long_value().unwrap();
-                                let activate = extract!(args[1], EncodableValue::Bool);
-
-                                let pointer = data.state.seat.get_pointer().unwrap();
-                                let keyboard = data.state.seat.get_keyboard().unwrap();
-
-                                let serial = SERIAL_COUNTER.next_serial();
-
-                                if !pointer.is_grabbed() {
-                                    let toplevel = data.state.xdg_toplevels.get(&(view_id as u64)).cloned();
-                                    if let Some(toplevel) = toplevel {
-                                        toplevel.with_pending_state(|state| {
-                                            if activate {
-                                                state.states.set(xdg_toplevel::State::Activated);
-                                            } else {
-                                                state.states.unset(xdg_toplevel::State::Activated);
-                                            }
-                                        });
-                                        keyboard.set_focus(&mut data.state, Some(toplevel.wl_surface().clone()), serial);
-
-                                        for toplevel in data.state.xdg_toplevels.values() {
-                                            toplevel.send_pending_configure();
-                                        }
-                                        result.success(None);
-                                    } else {
-                                        result.error(
-                                            "surface_doesnt_exist".to_string(),
-                                            format!("Surface {view_id} doesn't exist"),
-                                            None,
-                                        );
-                                    }
-                                }
-                            }
-                            _ => {
-                                result.success(None);
-                            }
-                        }
-                    }
-                },
-            )
-            .expect("Failed to init wayland server source");
-
         Self {
             running: Arc::new(AtomicBool::new(true)),
             display_handle,
@@ -337,7 +193,6 @@ impl<BackendData: Backend + 'static> ServerState<BackendData> {
             texture_ids_per_view_id: HashMap::new(),
             view_id_per_texture_id: HashMap::new(),
             texture_swapchains: HashMap::new(),
-            tx_platform_message: Some(tx_platform_message),
         }
     }
 }
@@ -549,7 +404,7 @@ impl<BackendData: Backend> CompositorHandler for ServerState<BackendData> {
 
                         Some(XdgPopupCommitMessage {
                             parent_id,
-                            geometry: dbg!(popup_data.current.geometry),
+                            geometry: popup_data.current.geometry,
                         })
                     }
                     _ => None,
